requirements.txt
datasets==3.2.0
matplotlib==3.10.0
numpy==2.2.1
Pillow==11.1.0
torch==2.5.1
torchvision==0.2.0
tqdm==4.67.1
